perf(factorization): narrow codes to u32, deterministic auto-dispatched parallel path
Implements the full factorization.rs code review: codes narrow to u32
throughout (representatives stay u64, and the factorize_arrow PyO3
boundary still widens codes back to u64 so downstream consumers that
require UInt64 arrays are unaffected). The generic hashable-key path
drops Option<K>-keyed hashing in favor of a separate null_code, and
null_count()==0 is specialized everywhere to skip per-row is_valid
checks and unblock autovectorization.

factorize_values_parallel is rewritten to be deterministic and
bit-for-bit identical to the sequential path (chunked single-hash
build, min-representative merge, first-seen code assignment by
sorting on representative), which lets parallel dispatch become
automatic by row count instead of a caller-supplied bool -- the
parallel_factorize/parallel knobs are removed from the Rust and
Python public surfaces.

Also: TotalF64 uses a monotonic bit-flip ordering key instead of
total_cmp per comparison; estimate_cardinality uses pseudo-random
sampling (not fixed-stride, which aliases on periodic data) with a
Chao1 extrapolation; sort_factorized partitions nulls before sorting
and parallelizes above a cardinality threshold; factorize_dense_integers
unifies its sort/no-sort branches and grows its span cap with input
size instead of hard-capping at 1M; factorize_boolean early-exits once
all 3 states are found; factorize_string_dictionary drops a dead pass
and an unneeded intermediate buffer; new dtype support for
Utf8View/BinaryView/Binary/LargeBinary and LargeUtf8-valued
dictionaries. factorize_bytes's string hasher was measured (not
assumed) against FxHash/foldhash/ahash on a 4M-row synthetic workload;
ahash won by ~5-12% and is now used there, with the reasoning and
numbers left as an inline comment.

+++ fastmob-core/src/measures/fitting/daily_location_lognormal.rs
@@ -9,8 +9,8 @@ use rustc_hash::{FxHashMap, FxHashSet};
 /// contains local calendar-day buckets, computed at the Arrow boundary after
 /// timezone metadata has been stripped.
 pub fn daily_unique_location_histogram_impl(
-    user_codes: &[u64],
-    location_codes: &[u64],
+    user_codes: &[u32],
+    location_codes: &[u32],
     days: &[i64],
     valid: &[bool],
 ) -> Result<(Vec<u64>, Vec<u64>), String> {
